+++ api/signup.go
@@ -384,6 +384,8 @@ func (a *Api) resendSignUp(res http.ResponseWriter, req *http.Request, vars map[
 
 			res.WriteHeader(http.StatusOK)
 		}
+	} else {
+		a.sendError(res, http.StatusNotFound, STATUS_SIGNUP_NOT_FOUND, "resendSignUp: sign up not found")
 	}
 }