+++ clients/smtpNotifier.go
@@ -22,7 +22,7 @@ type (
 	}
 )
 
-//NewSmtpNotifier creates a new SMTP notifier (using standard smtp to send emails)
+// NewSmtpNotifier creates a new SMTP notifier (using standard smtp to send emails)
 func NewSmtpNotifier(cfg *SmtpNotifierConfig) (*SmtpNotifier, error) {
 	return &SmtpNotifier{
 		Config: cfg,
@@ -39,9 +39,7 @@ func (c *SmtpNotifier) Send(to []string, subject string, msg string) (int, strin
 	}
 	mime := "MIME-version: 1.0;\nContent-Type: text/html; charset=\"UTF-8\";\n\n"
 	body := []byte("To: " + to[0] + "\r\n" +
-
 		"Subject: " + subject + "\r\n" +
-
 		mime + "\r\n" +
 		msg + "\r\n")
 	err := smtp.SendMail(c.Config.Server+":"+c.Config.Port, auth, c.Config.From, to, body)
@@ -49,5 +47,6 @@ func (c *SmtpNotifier) Send(to []string, subject string, msg string) (int, strin
 		log.Println(err.Error())
 		return 400, err.Error()
 	}
+	log.Printf("SMTP email sent: %s\n", subject)
 	return 200, "OK"
 }